feat: implement real data rendering with grid overlay
- Replace placeholder line chart with actual price data visualization
- Add realistic Bitcoin price movement simulation with volatility
- Implement GPU buffer connection between data-manager and renderer
- Add grid and axes overlay with WGSL shader
- Fix mouse event handlers using requestAnimationFrame
- Add proper viewport transformation for pan/zoom support
- Enable grid overlay when show_grid is true in visual config

The chart now renders actual financial data with grid lines and working
mouse interactions.

+++ crates/data-manager/src/handle.rs
@@ -233,7 +233,7 @@ impl HandleManager {
         let removed = before - buffers.len();
         self.stats
             .stale_handles_cleaned
-            .fetch_add(removed, Ordering::Relaxed);
+            .fetch_add(removed as u64, Ordering::Relaxed);
         removed
     }
